app/config/settings.py
import os
from dotenv import load_dotenv
from pydantic import BaseModel

# Load environment variables from .env file
load_dotenv()

class Settings(BaseModel):
    """Application settings loaded from environment variables."""
    DATABASE_URL: str = os.getenv("DATABASE_URL", "sqlite+aiosqlite:///./metadata.db")

    class Config:
        env_file = ".env"
        env_file_encoding = "utf-8"
        case_sensitive = False

# Create settings instance
settings = Settings() 