requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # For PostgreSQL
pymysql==1.1.0  # For MySQL
aiosqlite==0.19.0  # For SQLite 